put ensure function in conftest.py
