+++ tests/test_core.py
@@ -24,26 +24,10 @@ from subliminal.core import (
 from subliminal.subtitle import Subtitle
 from subliminal.utils import timestamp
 from subliminal.video import Episode, Movie
+from tests.conftest import ensure
 
 # Core test
 pytestmark = pytest.mark.core
-
-
-def ensure(path: str | os.PathLike[str], *, directory: bool = False) -> Path:
-    """Create a file (or directory) at path."""
-    path = Path(path)
-    if directory:
-        # Create a directory at path
-        if not path.is_dir():
-            path.mkdir(parents=True)
-
-    else:
-        # Create a directory at parent
-        if not path.parent.is_dir():
-            path.parent.mkdir(parents=True)
-        # Create a file at path
-        path.touch()
-    return path
 
 
 def test_check_video_languages(movies: dict[str, Movie]) -> None:
